# service/swagger/ducklake-api.yaml
openapi: 3.0.3
info:
  title: DuckLake Data Lakehouse API
  description: |
    REST API for DuckLake - Git-like branching for data lakehouses.

    ## Overview
    This API provides endpoints to:
    - Execute SQL queries and return tabular results
    - Manage data branches (create, delete, search, switch)
    - Query branch metadata and statistics
    - Perform time-travel queries
    - Connect to multiple catalog types (DuckLake, Databricks, Snowflake)

    ## Catalog Types
    The service supports multiple catalog types:
    - **DUCKLAKE**: Native DuckLake catalogs with full read/write access and branching support
    - **DATABRICKS**: Databricks Unity Catalog (read-only, uses OAuth service principal)
    - **SNOWFLAKE**: Snowflake database (read-only, not yet implemented)

    ## Authentication
    All endpoints require Bearer token authentication.

    ## Error Handling
    Errors follow Google's gRPC status code conventions.

  version: 1.0.0
  contact:
    name: DuckLake Team
    email: support@ducklake.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.ducklake.io/v1
    description: Production server
  - url: https://api.staging.ducklake.io/v1
    description: Staging server
  - url: http://localhost:8080/v1
    description: Local development
...
security:
  - BearerAuth: []
paths:
...
  /catalogs:
    get:
      tags:
        - Catalogs
      summary: List registered catalogs
      description: Returns a list of all catalogs registered in the service.
      operationId: listCatalogs
      parameters:
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/PageNumber'
        - name: metadata_type
          in: query
          description: Filter by metadata backend type
          schema:
            type: string
            enum: [postgres, duckdb, sqlite]
        - name: storage_type
          in: query
          description: Filter by storage type
          schema:
            type: string
            enum: [local, s3, gcs, azure]
        - name: catalog_type
          in: query
          description: Filter by catalog type
          schema:
            type: string
            enum: [DUCKLAKE, DATABRICKS, SNOWFLAKE]
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListCatalogsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Catalogs
      summary: Register catalog
      description: |
        Registers a new catalog in the service registry.

        The catalog is not immediately attached - it's stored in the service's
        metadata database and attached on-demand when queries reference it.

        ## Catalog Types

        **DUCKLAKE (default)** - Native DuckLake with full read/write and branching support:
        - Required fields: `metadata_uri`, `data_path`
        - Optional fields: `metadata_type`, `storage_type`, `secret_name`
        - Metadata backends: PostgreSQL, DuckDB file, SQLite
        - Data storage: Local, S3, GCS, Azure

        **DATABRICKS** - Databricks Unity Catalog (read-only):
        - Required fields: `databricks_workspace_url`, `databricks_client_id`,
          `databricks_client_secret`, `databricks_catalog_name`
        - Uses OAuth 2.0 client credentials flow for authentication

        **SNOWFLAKE** - Snowflake database (read-only, not yet implemented):
        - Required fields: `snowflake_account`, `snowflake_user`,
          `snowflake_password`, `snowflake_database`
        - Optional fields: `snowflake_warehouse`
      operationId: registerCatalog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterCatalogRequest'
      responses:
        '201':
          description: Catalog registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Catalog'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /catalogs/{catalog_id}:
    get:
      tags:
        - Catalogs
      summary: Get catalog
      description: Returns details of a registered catalog.
      operationId: getCatalog
      parameters:
        - $ref: '#/components/parameters/CatalogId'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Catalog'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    patch:
      tags:
        - Catalogs
      summary: Update catalog
      description: Updates a registered catalog's configuration.
      operationId: updateCatalog
      parameters:
        - $ref: '#/components/parameters/CatalogId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCatalogRequest'
      responses:
        '200':
          description: Catalog updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Catalog'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Catalogs
      summary: Unregister catalog
      description: |
        Removes a catalog from the service registry.
        Does not delete any data in the catalog itself.
      operationId: unregisterCatalog
      parameters:
        - $ref: '#/components/parameters/CatalogId'
      responses:
        '204':
          description: Catalog unregistered successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /catalogs/{catalog_id}:test:
    post:
      tags:
        - Catalogs
      summary: Test catalog connection
      description: |
        Tests connectivity to the catalog's metadata backend and data storage.
        Returns detailed status of each component.
      operationId: testCatalogConnection
      parameters:
        - $ref: '#/components/parameters/CatalogId'
      responses:
        '200':
          description: Connection test completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogTestResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ============================================================================
  # CATALOG DISCOVERY (Schemas/Tables)
  # ============================================================================
  /catalogs/{catalog_id}/schemas:
    get:
      tags:
        - Catalogs
      summary: List schemas
      description: Returns all schemas in the catalog visible from the current branch.
      operationId: listSchemas
      parameters:
        - $ref: '#/components/parameters/CatalogId'
        - name: branch
          in: query
          description: Branch to list schemas from (default current branch)
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListSchemasResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /catalogs/{catalog_id}/schemas/{schema_name}/tables:
    get:
      tags:
        - Catalogs
      summary: List tables
      description: Returns all tables in the schema visible from the current branch.
      operationId: listTables
      parameters:
        - $ref: '#/components/parameters/CatalogId'
        - $ref: '#/components/parameters/SchemaName'
        - name: branch
          in: query
          description: Branch to list tables from (default current branch)
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListTablesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /catalogs/{catalog_id}/schemas/{schema_name}/tables/{table_name}:
    get:
      tags:
        - Catalogs
      summary: Get table info
      description: Returns detailed information about a table including columns and statistics.
      operationId: getTable
      parameters:
        - $ref: '#/components/parameters/CatalogId'
        - $ref: '#/components/parameters/SchemaName'
        - $ref: '#/components/parameters/TableName'
        - name: branch
          in: query
          description: Branch to get table from (default current branch)
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableInfo'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ============================================================================
  # BRANCHES (Per-Catalog)
  # ============================================================================
...
  /catalogs/{catalog_id}/current-branch:
    get:
      tags:
        - Branches
      summary: Get current branch
      description: Returns the currently active working branch.
      operationId: getCurrentBranch
      parameters:
        - $ref: '#/components/parameters/CatalogId'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentBranchResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

# ============================================================================
# COMPONENTS
# ============================================================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication

  # --------------------------------------------------------------------------
  # PARAMETERS
  # --------------------------------------------------------------------------
  parameters:
    CatalogId:
      name: catalog_id
      in: path
      required: true
      description: Unique identifier for the DuckLake catalog
      schema:
        type: string
        example: "my_lakehouse"
...
    SchemaName:
      name: schema_name
      in: path
      required: true
      description: Name of the schema
      schema:
        type: string
        example: "main"
...
    PageSize:
      name: page_size
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    PageNumber:
      name: page_number
      in: query
      description: Page number (starts from 1)
      schema:
        type: integer
        minimum: 1
        default: 1

  # --------------------------------------------------------------------------
  # SCHEMAS
  # --------------------------------------------------------------------------
  schemas:
...
    RegisterCatalogRequest:
      type: object
      required:
        - catalog_id
      properties:
        catalog_id:
          type: string
          maxLength: 255
          pattern: '^[a-zA-Z][a-zA-Z0-9_]*$'
          description: Unique identifier for the catalog (alias name)
          example: "my_lakehouse"
        display_name:
          type: string
          maxLength: 255
          description: Human-readable display name
          example: "XVA Trading Desk"
        description:
          type: string
          maxLength: 1000
          description: Description of the catalog
          example: "Production XVA risk calculations and trade data"
        catalog_type:
          type: string
          enum: [DUCKLAKE, DATABRICKS, SNOWFLAKE]
          default: DUCKLAKE
          description: |
            Type of catalog to register:
            - DUCKLAKE: Native DuckLake with full read/write and branching support (default)
            - DATABRICKS: Databricks Unity Catalog (read-only)
            - SNOWFLAKE: Snowflake database (read-only, not yet implemented)
          example: "DUCKLAKE"
        # DuckLake-specific fields (required when catalog_type=DUCKLAKE)
        metadata_uri:
          type: string
          description: |
            Metadata backend connection string (required for DUCKLAKE):
            - PostgreSQL: `postgres:dbname=mydb host=localhost port=5432 user=user password=pass`
            - DuckDB file: `myfile.ducklake`
            - SQLite: `sqlite:myfile.db`
          example: "postgres:dbname=xva_desk host=localhost port=5433"
        metadata_type:
          type: string
          enum: [postgres, duckdb, sqlite]
          description: Type of metadata backend (DUCKLAKE only)
          example: "postgres"
        data_path:
          type: string
          description: |
            Data storage path (required for DUCKLAKE):
            - Local: `/path/to/data`
            - S3: `s3://bucket/path`
            - GCS: `gs://bucket/path`
            - Azure: `az://container/path`
          example: "/tmp/xva_desk"
        storage_type:
          type: string
          enum: [local, s3, gcs, azure]
          description: Type of data storage (DUCKLAKE only)
          example: "local"
        secret_name:
          type: string
          description: Name of secret for storage credentials (DUCKLAKE only)
          example: "my_s3_secret"
        # Databricks-specific fields (required when catalog_type=DATABRICKS)
        databricks_workspace_url:
          type: string
          description: Databricks workspace URL (required for DATABRICKS)
          example: "https://dbc-xxxxx.cloud.databricks.com"
        databricks_client_id:
          type: string
          description: Service principal client ID for OAuth (required for DATABRICKS)
          example: "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
        databricks_client_secret:
          type: string
          format: password
          description: Service principal client secret for OAuth (required for DATABRICKS)
          example: "dosexxxxxxxxxxxxxxxxxxxxxxx"
        databricks_catalog_name:
          type: string
          description: Unity Catalog name (required for DATABRICKS)
          example: "my_unity_catalog"
        # Snowflake-specific fields (required when catalog_type=SNOWFLAKE)
        snowflake_account:
          type: string
          description: Snowflake account identifier (required for SNOWFLAKE)
          example: "xy12345.us-east-1"
        snowflake_user:
          type: string
          description: Snowflake username (required for SNOWFLAKE)
          example: "my_user"
        snowflake_password:
          type: string
          format: password
          description: Snowflake password (required for SNOWFLAKE)
          example: "my_password"
        snowflake_database:
          type: string
          description: Snowflake database name (required for SNOWFLAKE)
          example: "my_database"
        snowflake_warehouse:
          type: string
          description: Snowflake warehouse name (optional for SNOWFLAKE)
          example: "compute_wh"
        # Common fields
        options:
          type: object
          additionalProperties:
            type: string
          description: Additional catalog options
          example:
            read_only: "false"
        tags:
          type: array
          items:
            type: string
          description: Tags for categorization
          example: ["production", "trading", "xva"]
...
    ListCatalogsResponse:
      type: object
      properties:
        catalogs:
          type: array
          items:
            $ref: '#/components/schemas/Catalog'
        page:
          $ref: '#/components/schemas/Page'

    # Schema/Table
...
    ListSchemasResponse:
      type: object
      properties:
        schemas:
          type: array
          items:
            $ref: '#/components/schemas/SchemaInfo'
        page:
          $ref: '#/components/schemas/Page'
...
  responses:
    BadRequest:
      description: Invalid request parameters (gRPC code 3 - INVALID_ARGUMENT)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Status'
          examples:
            validationFailed:
              summary: Validation failed
              value:
                code: 3
                message: "Validation failed"
                details:
                  - '@type': "type.googleapis.com/google.rpc.BadRequest"
                    field_violations:
                      - field: "branch_name"
                        description: "Branch name is required"
                      - field: "branch_name"
                        description: "Branch name must start with a letter"
            invalidSQL:
              summary: Invalid SQL syntax
              value:
                code: 3
                message: "SQL syntax error"
                details:
                  - '@type': "type.googleapis.com/google.rpc.ErrorInfo"
                    reason: "INVALID_SQL"
                    domain: "ducklake.io"
                    metadata:
                      position: "15"
                      expected: "FROM"
    Unauthorized:
      description: Missing or invalid authentication (gRPC code 16 - UNAUTHENTICATED)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Status'
          examples:
            unauthenticated:
              summary: No token provided
              value:
                code: 16
                message: "Authentication required"
                details:
                  - '@type': "type.googleapis.com/google.rpc.ErrorInfo"
                    reason: "CREDENTIALS_MISSING"
                    domain: "ducklake.io"
            tokenExpired:
              summary: Token expired
              value:
                code: 16
                message: "Token has expired"
                details:
                  - '@type': "type.googleapis.com/google.rpc.ErrorInfo"
                    reason: "TOKEN_EXPIRED"
                    domain: "ducklake.io"
...
    Conflict:
      description: Resource conflict (gRPC code 6 - ALREADY_EXISTS)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Status'
          examples:
            branchAlreadyExists:
              summary: Branch already exists
              value:
                code: 6
                message: "Branch 'feature_analytics' already exists"
                details:
                  - '@type': "type.googleapis.com/google.rpc.ResourceInfo"
                    resource_type: "Branch"
                    resource_name: "feature_analytics"
                    owner: "ducklake-service"
                    description: "A branch with this name already exists"
                  - '@type': "type.googleapis.com/google.rpc.ErrorInfo"
                    reason: "BRANCH_ALREADY_EXISTS"
                    domain: "ducklake.io"
                    metadata:
                      existing_branch_id: "3"
            cannotDeleteMainBranch:
              summary: Cannot delete main branch
              value:
                code: 9
                message: "Cannot delete main branch"
                details:
                  - '@type': "type.googleapis.com/google.rpc.ErrorInfo"
                    reason: "CANNOT_DELETE_MAIN_BRANCH"
                    domain: "ducklake.io"
            branchHasChildren:
              summary: Branch has active children
              value:
                code: 9
                message: "Cannot delete branch with active children"
                details:
                  - '@type': "type.googleapis.com/google.rpc.ErrorInfo"
                    reason: "BRANCH_HAS_CHILDREN"
                    domain: "ducklake.io"
                    metadata:
                      child_count: "3"
    InternalServerError:
      description: Internal server error (gRPC code 13 - INTERNAL)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Status'
          examples:
            internalError:
              summary: Internal error
              value:
                code: 13
                message: "Internal server error"
                details:
                  - '@type': "type.googleapis.com/google.rpc.ErrorInfo"
                    reason: "INTERNAL_ERROR"
                    domain: "ducklake.io"
            metadataConnectionError:
              summary: Metadata database connection error
              value:
                code: 14
                message: "Unable to connect to metadata database"
                details:
                  - '@type': "type.googleapis.com/google.rpc.ErrorInfo"
                    reason: "METADATA_CONNECTION_ERROR"
                    domain: "ducklake.io"
                    metadata:
                      retry_after: "30"
            storageError:
              summary: Storage access error
              value:
                code: 13
                message: "Unable to access data storage"
                details:
                  - '@type': "type.googleapis.com/google.rpc.ErrorInfo"
                    reason: "STORAGE_ERROR"
                    domain: "ducklake.io"